% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_chunk.R
\name{splitChunk}
\alias{splitChunk}
\title{Split code chunk.}
\usage{
splitChunk()
}
\value{
Inserts ```\\n\\n```\{r\}
}
\description{
Addin for RStudio for splitting code chunks in R Markdown.

See \code{details} for setting up key command.
}
\details{
How to set up key command in RStudio:

After installing package.
Go to Tools >> Addins >> Browse Addins >> Keyboard Shortcuts.
Find "Split Code Chunk" and press its field under Shortcut.
Press desired key command.
Press Apply.
Press Execute.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}

