% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_chunk.R
\name{splitChunk}
\alias{splitChunk}
\title{Split code chunk.}
\usage{
splitChunk()
}
\value{
Inserts ``` \\n\\n ```\{r\}
}
\description{
Addin for RStudio for splitting code chunks in RMarkdown.

See \code{details} for setting up keycommand.
}
\details{
How to set up key command:
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}

